feat: move files, creating VideoContinueModal
