+++ src/store/useLectureStore.ts
@@ -13,3 +13,17 @@ export const useLectureStore = create<LectureStore>((set) => ({
   setSelectedChapterId: (id) => set({selectedChapterId: id}),
   setSelectedVideoId: (id) => set({selectedVideoId: id}),
 }));
+
+type VideoStore = {
+  playing: boolean;
+  duration: number;
+  setPlaying: (state: boolean) => void;
+  setDuration: (total: number) => void;
+}
+
+export const useVideoStore = create<VideoStore>((set) => ({
+  playing: false,
+  duration: 0,
+  setPlaying: (state) => set({playing: state}),
+  setDuration: (total) => set({duration: total}),
+}))
